refactor(task-session-manager): extract tool hooks, evaluator, event router
Final extraction phase. index.ts thinned from 881 to 299 lines by moving:
- tool.execute.before/after -> tool-execute-hooks.ts
- evaluateContinuation god-function -> continuation-evaluator.ts
  (deduped the repeated 5-condition guard into isEvaluationAborted)
- event handler switch -> event-router.ts

All handlers take explicit deps; no behavior change. Cache-safety
allowlist updated for the Date.now() call now located in event-router.ts.
